refactor(plugin): decompose 1424L manager.go god file into focused modules
Split internal/plugin/manager.go into six single-responsibility files:
- installer.go: Install/Remove/Update + lock serialization + reverse-deps
- download.go:  downloadPlugin/extractTarGz/rollbackInstall
- loader.go:    ListInstalled/LoadManifestsFromDir/List + nginx route scan
- security.go:  checksum/Ed25519 signature/CRL/license/EOL/postInstallEvent
- config.go:    DisablePlugin/EnablePlugin/IsDisabled + table-conflict guards
- paths.go:     DefaultCacheDir/LicenseCacheDir/pingAPIURL/findPlugin wrappers

manager.go reduced to type declarations only (<40L). All exported symbols
preserved — no callers require changes. go build ./... passes clean.

Ticket: T-E2-04 (Mega Phase 1, Epic E2)
